move tests out of contact form and into new test file for email renderer
